+++ packages/cube/test/validation.pipeline.test.js
@@ -15,7 +15,7 @@ describe('cube validation pipeline', function () {
     const result = shell.exec(command, { silent: true, cwd })
 
     strictEqual(result.stderr, '')
-    strictEqual(result.stdout, '')
+    ok(result.stdout.includes('_:report <http://www.w3.org/ns/shacl#conforms> "true"^^<http://www.w3.org/2001/XMLSchema#boolean>'))
     strictEqual(result.code, 0)
   })
 
@@ -28,6 +28,19 @@ describe('cube validation pipeline', function () {
     strictEqual(result.code, 1)
     expect(result.stderr).to.match(/1 violations found/)
     ok(result.stdout.includes('_:report <http://www.w3.org/ns/shacl#conforms> "false"^^<http://www.w3.org/2001/XMLSchema#boolean>'))
+  })
+
+  it('should run check-cube-observations pipeline with warning', () => {
+    const constraintFile = `${support}/constraint01.ttl`
+    const command = `cat ${support}/observations03.ttl | barnard59 cube check-observations --constraint ${constraintFile}`
+
+    const result = shell.exec(command, { silent: true, cwd })
+
+    strictEqual(result.code, 0) // successful because there are no reults with severity sh:Violation
+    expect(result.stderr).to.match(/1 results with severity http:\/\/www.w3.org\/ns\/shacl#Warning/)
+    // sh:conforms is false because of the warning
+    strictEqual(true, result.stdout.includes('_:report <http://www.w3.org/ns/shacl#conforms> "false"^^<http://www.w3.org/2001/XMLSchema#boolean>'))
+    strictEqual(false, result.stdout.includes('_:report <http://www.w3.org/ns/shacl#conforms> "true"^^<http://www.w3.org/2001/XMLSchema#boolean>'))
   })
 
   it('should run check-cube-observations when maxViolations is not exceeded', () => {
@@ -48,7 +61,31 @@ describe('cube validation pipeline', function () {
     const result = shell.exec(command, { silent: true, cwd })
 
     strictEqual(result.stderr, '')
-    strictEqual(result.stdout, '')
+    ok(result.stdout.includes('_:report <http://www.w3.org/ns/shacl#conforms> "true"^^<http://www.w3.org/2001/XMLSchema#boolean>'))
+    strictEqual(result.code, 0)
+  })
+
+  it('should run check-cube-observations pipeline with error for cube.link examples', () => {
+    const constraintFile = `${support}/constraint.ttl`
+    // sh:class constraint fails because of batching
+    const command = `cat ${support}/cube.ttl | barnard59 cube check-observations --constraint ${constraintFile} --batchSize 10`
+
+    const result = shell.exec(command, { silent: true, cwd })
+
+    strictEqual(result.code, 1)
+    ok(result.stdout.includes('_:report <http://www.w3.org/ns/shacl#conforms> "false"^^<http://www.w3.org/2001/XMLSchema#boolean>'))
+  })
+
+  it('should run check-cube-observations pipeline with cube.link examples', () => {
+    const constraintFile = `${support}/constraint.ttl`
+    // disabling batching makes all data available when checking the sh:class constraint
+    const command = `cat ${support}/cube.ttl | barnard59 cube check-observations --constraint ${constraintFile} --batchSize 0`
+
+    const result = shell.exec(command, { silent: true, cwd })
+
+    strictEqual(result.stderr, '')
+    expect(result.stdout)
+      .to.include('_:report <http://www.w3.org/ns/shacl#conforms> "true"^^<http://www.w3.org/2001/XMLSchema#boolean>')
     strictEqual(result.code, 0)
   })
 })