+++ packages/formats/csvw.js
@@ -1,15 +1,50 @@
 import sinkToDuplex from '@rdfjs/sink-to-duplex'
-import CsvwParser from 'rdf-parser-csvw'
+import CsvwParser from '@zazuko/rdf-parser-csvw'
 import tracer from './lib/tracer.js'
 import { toDataset } from './lib/stream.js'
 
+/**
+ * @typedef {{
+ *   metadata: import('@rdfjs/types').DatasetCore
+ *   timezone?: string
+ *   relaxColumnCount?: boolean
+ *   skipLinesWithError?: boolean
+ *   trimHeaders?: boolean
+ *   skipEmptyLines?: boolean
+ * }} Options
+ */
+
+/**
+ * @overload
+ * @param {Options} args
+ * @return {import('duplexify').Duplexify}
+ */
+
+/**
+ * @overload
+ * @param {import('@rdfjs/types').DatasetCore} metadata
+ * @return {import('duplexify').Duplexify}
+ */
+
+/**
+ * @this {import('barnard59-core').Context}
+ * @param {Options | import('@rdfjs/types').DatasetCore} args
+ */
 function parse(args) {
+  /**
+   * @type {import('@rdfjs/types').DatasetCore}
+   */
   let metadata
   let relaxColumnCount = false
   let skipLinesWithError = false
   let timezone = 'local'
+  let trimHeaders = true
+  /**
+   * @type {Boolean | undefined}
+   */
+  let skipEmptyLines
 
-  if (args.metadata) {
+  if ('metadata' in args) {
     metadata = args.metadata
 
     if (typeof args.relaxColumnCount !== 'undefined') {
@@ -23,6 +58,14 @@ function parse(args) {
     if (typeof args.timezone !== 'undefined') {
       timezone = args.timezone
     }
+
+    if (typeof args.trimHeaders !== 'undefined') {
+      trimHeaders = Boolean(args.trimHeaders)
+    }
+
+    if (typeof args.skipEmptyLines !== 'undefined') {
+      skipEmptyLines = Boolean(args.skipEmptyLines)
+    }
   } else {
     metadata = args
   }
@@ -32,10 +75,13 @@ function parse(args) {
       const dataset = await toDataset(this.env, metadata)
       span.addEvent('metadata')
       return sinkToDuplex(new CsvwParser({
+        factory: this.env,
         metadata: dataset,
         relaxColumnCount,
         skipLinesWithError,
         timezone,
+        trimHeaders,
+        skipEmptyLines,
       }), {
         readableObjectMode: true,
       })