+++ packages/formats/rdf-xml.js
@@ -1,8 +1,13 @@
 import sinkToDuplex from '@rdfjs/sink-to-duplex'
 import { RdfXmlParser } from 'rdfxml-streaming-parser'
 
+/**
+ * @this {import('barnard59-core').Context}
+ */
 function parse() {
-  return sinkToDuplex(new RdfXmlParser(), {
+  return sinkToDuplex(new RdfXmlParser({
+    dataFactory: this.env,
+  }), {
     readableObjectMode: true,
     writableObjectMode: true,
   })