+++ packages/ftp/test/write.test.js
@@ -3,7 +3,7 @@ import fs from 'node:fs'
 import { dirname, resolve } from 'node:path'
 import write from '../write.js'
 import { withServer } from './support/server.js'
-import { ftpConfigurations } from './support/serverConfigurations.js'
+import ftpConfigurations from './support/serverConfigurations.js'
 import { finished } from './support/stream.js'
 
 export const __dirname = dirname(new URL(import.meta.url).pathname)
@@ -23,6 +23,12 @@ describe('write', () => {
         })
 
         strictEqual(fs.readFileSync(resolve(__dirname, 'support/tmp/foo.txt')).toString(), '987\n654')
+      })
+
+      afterEach(() => {
+        if (fs.existsSync(resolve(__dirname, 'support/tmp/foo.txt'))) {
+          fs.rmSync(resolve(__dirname, 'support/tmp/foo.txt'))
+        }
       })
     })
 })